Login home 03 07 (#1)
* Update HomeView.swift

After user logging in, we are redirected to the home page, where we have two options: start a new match or view the results of completed matches.

* Update SignUpView.swift

A new user must sign up before using the app. The sign-up page includes two fields: username and password, allowing users to register. These credentials can then be used to log in to the app.